[設定] b config 分檔
