+++ stubs/section/b/corePlugins.js
@@ -1,0 +1,52 @@
+module.exports = {
+  accessibility: false,
+  appearance: false,
+  backgroundClip: false,
+  backgroundImage: false,
+  backgroundOpacity: false,
+  borderOpacity: false,
+  container: false,
+  divideColor: false,
+  divideOpacity: false,
+  divideStyle: false,
+  divideWidth: false,
+  fontFamily: false,
+  fontSmoothing: false,
+  fontVariantNumeric: false,
+  gap: false,
+  gradientColorStops: false,
+  gridAutoColumns: false,
+  gridAutoFlow: false,
+  gridAutoRows: false,
+  gridColumn: false,
+  gridColumnEnd: false,
+  gridColumnStart: false,
+  gridRow: false,
+  gridRowEnd: false,
+  gridRowStart: false,
+  gridTemplateColumns: false,
+  gridTemplateRows: false,
+  letterSpacing: false,
+  objectFit: false,
+  objectPosition: false,
+  overscrollBehavior: false,
+  placeContent: false,
+  placeholderOpacity: false,
+  placeItems: false,
+  placeSelf: false,
+  rotate: false,
+  scale: false,
+  skew: false,
+  space: false,
+  textOpacity: false,
+  transform: false,
+  transformOrigin: false,
+  transitionDelay: false,
+  transitionDuration: false,
+  transitionProperty: false,
+  transitionTimingFunction: false,
+  translate: false,
+
+  // 禁用 c guideline 相關
+  cColor: false
+}
