+++ stubs/section/b/variants.js
@@ -1,0 +1,114 @@
+module.exports = {
+  accessibility: ['responsive', 'focus'],
+  alignContent: ['responsive'],
+  alignItems: ['responsive'],
+  alignSelf: ['responsive'],
+  appearance: ['responsive'],
+  backgroundAttachment: ['responsive'],
+  backgroundClip: ['responsive'],
+  backgroundColor: ['responsive', 'hover', 'focus'],
+  backgroundImage: ['responsive'],
+  gradientColorStops: ['responsive', 'hover', 'focus'],
+  backgroundOpacity: ['responsive', 'hover', 'focus'],
+  backgroundPosition: ['responsive'],
+  backgroundRepeat: ['responsive'],
+  backgroundSize: ['responsive'],
+  borderCollapse: ['responsive'],
+  borderColor: ['responsive', 'hover', 'focus'],
+  borderOpacity: ['responsive', 'hover', 'focus'],
+  borderRadius: ['responsive'],
+  borderStyle: ['responsive'],
+  borderWidth: ['responsive'],
+  boxShadow: ['responsive', 'hover', 'focus'],
+  boxSizing: ['responsive'],
+  container: ['responsive'],
+  cursor: ['responsive'],
+  display: ['responsive'],
+  divideColor: ['responsive'],
+  divideOpacity: ['responsive'],
+  divideStyle: ['responsive'],
+  divideWidth: ['responsive'],
+  fill: [],
+  flex: ['responsive'],
+  flexDirection: ['responsive'],
+  flexGrow: ['responsive'],
+  flexShrink: ['responsive'],
+  flexWrap: ['responsive'],
+  float: ['responsive'],
+  clear: ['responsive'],
+  fontFamily: ['responsive'],
+  fontSize: ['responsive'],
+  fontSmoothing: ['responsive'],
+  fontVariantNumeric: ['responsive'],
+  fontStyle: ['responsive'],
+  fontWeight: ['responsive', 'hover', 'focus'],
+  height: ['responsive'],
+  inset: ['responsive'],
+  justifyContent: ['responsive'],
+  justifyItems: ['responsive'],
+  justifySelf: ['responsive'],
+  letterSpacing: ['responsive'],
+  lineHeight: ['responsive'],
+  listStylePosition: ['responsive'],
+  listStyleType: ['responsive'],
+  margin: ['responsive'],
+  maxHeight: ['responsive'],
+  maxWidth: ['responsive'],
+  minHeight: ['responsive'],
+  minWidth: ['responsive'],
+  objectFit: ['responsive'],
+  objectPosition: ['responsive'],
+  opacity: ['responsive', 'hover', 'focus'],
+  order: ['responsive'],
+  outline: ['responsive', 'focus'],
+  overflow: ['responsive'],
+  overscrollBehavior: ['responsive'],
+  padding: ['responsive'],
+  placeContent: ['responsive'],
+  placeItems: ['responsive'],
+  placeSelf: ['responsive'],
+  placeholderColor: ['responsive', 'focus'],
+  placeholderOpacity: ['responsive', 'focus'],
+  pointerEvents: ['responsive'],
+  position: ['responsive'],
+  resize: ['responsive'],
+  space: ['responsive'],
+  stroke: [],
+  strokeWidth: [],
+  tableLayout: ['responsive'],
+  textAlign: ['responsive'],
+  textColor: ['responsive', 'hover', 'focus'],
+  textOpacity: ['responsive', 'hover', 'focus'],
+  textDecoration: ['responsive', 'hover', 'focus'],
+  textTransform: ['responsive'],
+  userSelect: ['responsive'],
+  verticalAlign: ['responsive'],
+  visibility: ['responsive'],
+  whitespace: ['responsive'],
+  width: ['responsive'],
+  wordBreak: ['responsive'],
+  zIndex: ['responsive'],
+  gap: ['responsive'],
+  gridAutoFlow: ['responsive'],
+  gridTemplateColumns: ['responsive'],
+  gridAutoColumns: ['responsive'],
+  gridColumn: ['responsive'],
+  gridColumnStart: ['responsive'],
+  gridColumnEnd: ['responsive'],
+  gridTemplateRows: ['responsive'],
+  gridAutoRows: ['responsive'],
+  gridRow: ['responsive'],
+  gridRowStart: ['responsive'],
+  gridRowEnd: ['responsive'],
+  transform: ['responsive'],
+  transformOrigin: ['responsive'],
+  scale: ['responsive', 'hover', 'focus'],
+  rotate: ['responsive', 'hover', 'focus'],
+  translate: ['responsive', 'hover', 'focus'],
+  skew: ['responsive', 'hover', 'focus'],
+  transitionProperty: ['responsive'],
+  transitionTimingFunction: ['responsive'],
+  transitionDuration: ['responsive'],
+  transitionDelay: ['responsive'],
+  animation: ['responsive'],
+}
